requirements.txt
neo4j==5.15.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
kaggle==1.5.16

